requirements.txt
# FastAPI and API dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart==0.0.9

# Data Science and ML - Versiones específicas para V2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
scipy==1.12.0

# Time Series Forecasting - V2 específico
statsmodels==0.14.1
arch==6.3.0

# Advanced ML Models - V2 optimizado
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.6.1

# Feature Engineering V2
tqdm==4.66.2

# Cache and Redis
redis==5.0.1

# Feature Engineering
# feature-engine>=1.6.0  # Uncomment if needed
# tsfresh>=0.20.0  # Uncomment if needed

# HTTP Clients for APIs
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Financial Data APIs
yfinance>=0.2.20
# quandl>=3.7.0  # Uncomment if using Quandl
# fredapi>=0.4.3  # Uncomment if using FRED API

# Database and Caching
sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0  # Uncomment for PostgreSQL
alembic>=1.12.0
# redis>=5.0.0  # Uncomment for Redis caching
cachetools>=5.3.0

# Environment and Configuration
python-dotenv>=1.0.0

# Logging and Monitoring
loguru>=0.7.0
# prometheus-client>=0.18.0  # Uncomment for metrics

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Jupyter for analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3
